+++ yaralyzer/output/decoding_attempts_table.py
@@ -25,7 +25,7 @@ from rich.text import Text
 
 from yaralyzer.bytes_match import BytesMatch
 from yaralyzer.util.helpers.bytes_helper import ascii_view_of_raw_bytes, hex_view_of_raw_bytes, rich_text_view_of_raw_bytes
-from yaralyzer.util.helpers.rich_text_helper import CENTER, FOLD, MIDDLE, RIGHT, na_txt
+from yaralyzer.util.helpers.rich_helper import CENTER, FOLD, MIDDLE, RIGHT, na_txt
 
 HEX = Text('HEX', style='bytes.title')
 RAW_BYTES = Text('Raw', style=f"bytes")