+++ yaralyzer/yara/yara_match.py
@@ -29,7 +29,7 @@ from rich.text import Text
 from yara import StringMatch
 
 from yaralyzer.util.helpers.bytes_helper import clean_byte_string
-from yaralyzer.util.helpers.rich_text_helper import CENTER
+from yaralyzer.util.helpers.rich_helper import CENTER
 from yaralyzer.util.helpers.string_helper import INDENT_SPACES
 from yaralyzer.output.rich_console import console_width, theme_colors_with_prefix
 from yaralyzer.util.logging import log