+++ yaralyzer/yaralyzer.py
@@ -13,7 +13,7 @@ from yaralyzer.bytes_match import BytesMatch
 from yaralyzer.config import YaralyzerConfig
 from yaralyzer.decoding.bytes_decoder import BytesDecoder
 from yaralyzer.util.helpers.file_helper import files_in_dir, load_binary_data, to_paths
-from yaralyzer.util.helpers.rich_text_helper import dim_if, reverse_color, print_fatal_error_and_exit
+from yaralyzer.util.helpers.rich_helper import dim_if, reverse_color, print_fatal_error_and_exit
 from yaralyzer.util.helpers.string_helper import comma_join, newline_join
 from yaralyzer.output.file_hashes_table import bytes_hashes_table
 from yaralyzer.output.regex_match_metrics import RegexMatchMetrics